fix(control_server): fix control server 'wait_till_done' method
This commit makes sure that the right effort is compared when we check
whether the control effort has been reached. It also renames the
'wait_till_done' method to 'wait_till_arm_control_done'. Lastly it
comments out its use for effort control. This was done since the 'FrankaHWSim'
does not yet simulate control delays; as a result, torque control commands
are applied instantaneously and waiting is not needed for effort control.